nidas::dynld::raf::DSMArincSensor:     Added capability to apply a VariableConverter to any label.     This required changing the convert method of VariableConverter     to pass and return doubles, since some position variables are doubles     and not floats.
nidas::core::VariableConverter, Linear and Polynomial, nidas::dynld::isff::CS_Krypton:
    convert method now returns double

nidas::core::Variable:
    getConverter() is also now overloaded, returning a const or non-const
    VariableConverter* depending on the the const-ness of Variable.

    It was bad practice to return a non-const VariableConverter* from
    a const Variable.

    Could have gotten around this issue by making VariableConverter::convert() a const
    method, but a convert() can read a cal file, changing its state and
    changing the current coefficents of the conversion.

nidas::core::SampleTag:
    Overloaded getVariables() for a non-const SampleTag to return
    a vector of non-const pointers to Variables, to support the above
    change to Variable.

nidas::core::CharacterSensor
nidas::dynld::A2DSensor
nidas::dynld::raf::DSMAnalogSensor
nidas::dynld::raf::PSI9116_Sensor
nidas::dynld::isff::CSAT3_Sonic
nidas::dynld::isff::PropVane
nidas::dynld::isff::WisardMote
nidas::dynld::iss::WICORSensor
nidas::dynld::StatisticsCruncher
nidas::dynld::isff::NetcdfRPCChannel
    Upates for the changes to Variable, SampleTag and VariableConverter.

tests/sync_server_dump:
    Updates to test a cal file applied to LAT from raf.GPS_HW_HG2021GB02
    Right now the lat.dat cal file just applies a linear (0,1) conversion which
    amounts to no change to the results. Once the nidas trunk has been changed to
    apply conversions to aircraft variables, the coefficients can be changed to
    verify every time that they are actually being applied.
    Add a memory leak suppression for Process::setEnvVar().

+++ src/nidas/core/VariableConverter.cc
@@ -284,7 +284,7 @@ void Linear::readCalFile(dsm_time_t t)
     }
 }
 
-float Linear::convert(dsm_time_t t,float val)
+double Linear::convert(dsm_time_t t,double val)
 {
     readCalFile(t);
     return val * _slope + _intercept;
@@ -475,11 +475,12 @@ void Polynomial::readCalFile(dsm_time_t t)
     }
 }
 
-float Polynomial::convert(dsm_time_t t,float val)
+double Polynomial::convert(dsm_time_t t,double val)
 {
     readCalFile(t);
     return eval(val,_coefs,_ncoefs);
 }
+
 
 std::string Polynomial::toString()
 {
