+++ src/nidas/core/VariableConverter.h
@@ -57,7 +57,7 @@ public:
 
     virtual const CalFile* getCalFile() const = 0;
 
-    virtual float convert(dsm_time_t,float v) = 0;
+    virtual double convert(dsm_time_t,double v) = 0;
 
     void setUnits(const std::string& val) { _units = val; }
 
@@ -169,7 +169,7 @@ public:
 
     void readCalFile(dsm_time_t t);
 
-    float convert(dsm_time_t t,float val);
+    double convert(dsm_time_t t,double val);
 
     std::string toString();
 
@@ -234,7 +234,7 @@ public:
 
     void readCalFile(dsm_time_t t);
 
-    float convert(dsm_time_t t,float val);
+    double convert(dsm_time_t t,double val);
 
     std::string toString();
 
@@ -244,7 +244,7 @@ public:
     void fromDOMElement(const xercesc::DOMElement*)
     	throw(nidas::util::InvalidParameterException);
 
-    static float eval(float x,float *p, int np);
+    static double eval(double x,float *p, int np);
 
 private:
 
@@ -267,7 +267,7 @@ private:
 };
 
 /* static */
-inline float Polynomial::eval(float x,float *p, int np)
+inline double Polynomial::eval(double x,float *p, int np)
 {
     double y = 0.0;
     for (int i = np - 1; i > 0; i--) {
