+++ src/nidas/dynld/isff/CS_Krypton.h
@@ -109,7 +109,7 @@ public:
     /**
      * Convert a voltage to water vapor density in g/m^3.
      */
-    float convert(dsm_time_t t, float volts);
+    double convert(dsm_time_t t, double volts);
 
     std::string toString();
